# spec/finance_research_openapi.yaml
openapi: 3.1.0
info:
  title: Finance Research App API
  version: 1.0.0
  description: Core contracts for subjects, market, fundamentals, evidence, snapshots, chat, analyze, agents, home, watchlists, and screener. The /v1 API is the client-facing backend-for-frontend surface.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
# Contract rule: Clients must not call third-party providers directly.
...
paths:
...
  /v1/market/quote:
    get:
      tags: [Market]
      summary: Get latest quote for a subject
      operationId: getQuote
      parameters:
        - in: query
          name: subject_kind
          required: true
          schema:
            $ref: '#/components/schemas/SubjectKind'
        - in: query
          name: subject_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Latest quote facts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactListResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SubjectKind:
      type: string
      description: Canonical subject discriminator. Ticker is a listing locator, not canonical identity.
      enum: [issuer, instrument, listing, theme, macro_topic, portfolio, screen]
...
    Fact:
      type: object
      description: Facts are the unit of truth for displayed values.
      properties:
        fact_id:
          type: string
          format: uuid
        metric_key:
          type: string
        value_num:
          type: number
        value_text:
          type: string
        unit:
          type: string
        currency:
          type: string
        as_of:
          type: string
          format: date-time
        source_id:
          type: string
          format: uuid
        verification_status:
          type: string
        freshness_class:
          type: string
        coverage_level:
          type: string
    FactListResponse:
      type: object
      properties:
        facts:
          type: array
          items:
            $ref: '#/components/schemas/Fact'
...